django/requirements.txt
django >= 1.0,<2.0
# psycopg2 >= 2.0,<3.0
psycopg2-binary >= 2.0,<3.0
gunicorn == 19.9.0
awscli==1.16.230

djangorestframework>=3.9.1
django-cors-headers==2.4.0
djangorestframework-jwt >= 1.11,<1.12

django-filter==2.2.0
djangorestframework-filters>=1.0.0.dev0

rest-social-auth < 2.1

django-cacheops<=4.3

# REST Social Auth dependencies: https://github.com/st4lk/django-rest-social-auth
social-auth-core >=3.0, <4.0
social-auth-app-django >=3.1, <4.0

# For access to AWS S3
boto3==1.9.27

elasticsearch-dsl==7.1.0